BIGINT ga o'zgartirildi (foydalanuvchilar bazasi)

+++ PHP/8-dars/1-vazifa.sql
@@ -6,6 +6,6 @@ CREATE TABLE foydalanuvchilar (
     id INT AUTO_INCREMENT PRIMARY KEY,
     ism VARCHAR(25) NOT NULL,
     familiya VARCHAR(25) NOT NULL,
-    telefon_raqam INT,
+    telefon_raqam BIGINT(13),
     viloyat VARCHAR(35)
 );
